Add JavaDocs for all classes except usertask and utils pkgs
