+++ src/main/java/duke/storage/SaveHandler.java
@@ -9,6 +9,12 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Scanner;
 
+/**
+ * SaveHandler is the class that handles all storage-related information and methods of the Duke chat-bot.
+ * A SaveHandler object encapsulates the necessary information for storing and restoring of the user's saved tasks.
+ * This includes the directory which is being used as the storage, and some Enums being used to parse
+ * the saved strings.
+ */
 public class SaveHandler {
     private final static String DATA_DIRECTORY = "data";
     private final static String DATA_FILEPATH = DATA_DIRECTORY + "/duke.txt";
@@ -17,6 +23,14 @@ public class SaveHandler {
         T, D, E
     }
 
+    /**
+     * Creates a new SaveHandler object which is able to save and restore tasks to a pre-defined location.
+     * SaveHandler attempts to create the necessary directories and files without explicitly checking if they
+     * already exist. Since creating existing directories and files only fails silently, there is no explicit checks
+     * done.
+     *
+     * @throws IOException Thrown if the creation/opening of the files and directories failed.
+     */
     public SaveHandler() throws IOException {
         File dir = new File(DATA_DIRECTORY);
         dir.mkdirs();
@@ -24,6 +38,11 @@ public class SaveHandler {
         f.createNewFile();
     }
 
+    /**
+     * Saves the list of tasks specified to the storage. The existing storage data is overwritten.
+     *
+     * @param tasks List of tasks to save.
+     */
     public void save(TaskList tasks) {
         try {
             FileWriter fw = new FileWriter(DATA_FILEPATH);
@@ -34,6 +53,13 @@ public class SaveHandler {
         }
     }
 
+    /**
+     * Restores the list of tasks saved on a previous execution of the Duke chat-bot into a <code>TaskList</code>
+     * object.
+     *
+     * @param emptyTasks The <code>TaskList</code> object being maintained by Duke.
+     * @throws DukeException Thrown if the restoration failed.
+     */
     public void restore(TaskList emptyTasks) throws DukeException {
         try {
             File saveFile = new File(DATA_FILEPATH);
